+++ CHANGELOG.md
@@ -8,6 +8,10 @@ and this project adheres to
 
 ## [Unreleased]
 
+### Added
+
+- ✨(teams) allow team management for team admins/owners #509
+
 ## [1.5.0] - 2024-11-14
 
 ### Removed